fix: perbaiki type mismatch Long vs Double di PimpinanHeroSummaryCard
totalPiutang di AdminSummary bertipe Long, bukan Double.
- Ubah parameter PimpinanHeroSummaryCard: Double -> Long
- Hapus .toInt() pada formatRupiah(totalPiutang) karena Utils.kt
  sudah punya overload formatRupiah(Long)

+++ app/src/main/kotlin/com/example/koperasikitagodangulu/PimpinanReportsScreen.kt
@@ -926,7 +926,7 @@ private fun DaftarSemuaNasabahCardClickable(
 @Composable
 private fun PimpinanHeroSummaryCard(
     totalAktif: Int,
-    totalPiutang: Double,
+    totalPiutang: Long,
     jumlahPdl: Int,
     isDark: Boolean = false
 ) {
@@ -977,7 +977,7 @@ private fun PimpinanHeroSummaryCard(
                 )
                 Spacer(modifier = Modifier.height(4.dp))
                 Text(
-                    text = formatRupiah(totalPiutang.toInt()),
+                    text = formatRupiah(totalPiutang),
                     fontSize = 26.sp,
                     fontWeight = FontWeight.Bold,
                     color = Color.White
